+++ src/components/PathwaysList/PathwaysList.tsx
@@ -16,7 +16,7 @@ import PathwayModal from './PathwayModal';
 
 import useStyles from './styles';
 import FileImportModal from 'components/FileImportModal';
-import useListCheckbox from './PathwaysListCheckbox';
+import useListCheckbox from 'hooks/useListCheckbox';
 import { useCriteriaContext } from 'components/CriteriaProvider';
 import { downloadPathway } from 'utils/builder';
